Fix diff output
less -r was messing with the formatting where tabs were involved (ie not
showing them, so tidied output looked untidy)

Changing color from always to auto.  "The default setting is auto, which
colors output when it’s going straight to a terminal, but omits the
color-control codes when the output is redirected to a pipe or a file."

Output looked terrible when I piped a diff to a file.