Add a config file for bat
fzf.vim uses bat for previews. The default preview style uses a
background colour for selected lines that is really hard to view over the
background of Nord. This allows me to use Nord as my theme within bat,
making it much easier to spot selected lines in a fzf preview pane.